Remove `host_udf.hpp` dependency from `detail/aggregation/aggregation.hpp`
